refactor: refresh feature_stats_v1 with percentile stats and fix PostgreSQL NUMERIC type
- Add percentile fields (P10-P90) to FeatureStats for richer distribution modeling
- Fix AVG(BIGINT) NUMERIC type mismatch with explicit DOUBLE PRECISION casts
- Introduce DistributionStats, GlobalCollectionStats, LoopReferenceProfile structs
- Remove deprecated useless/ module (6.3K lines cleanup)
- Migrate from GifMeta to LoopMeta for consistent metadata handling
- Add refresh_stats binary for on-demand statistics recalculation
- Update CHANGELOG with v0.11.1 (2026-04-03) changes

+++ shared_utils/src/video_explorer.rs
@@ -2613,7 +2613,7 @@ impl VideoExplorer {
         let ms_ssim = if self.config.quality_thresholds.validate_ms_ssim {
             let duration = get_video_duration(&self.input_path);
             let ms_ssim_skip_threshold_secs = if self.config.ultimate_mode {
-                VMAF_SKIP_THRESHOLD_ULTIMATE_SECS
+                f64::from(VMAF_SKIP_THRESHOLD_ULTIMATE_SECS)
             } else {
                 f64::from(LONG_VIDEO_THRESHOLD_SECS)
             };
